+++ src/main/java/com/diffplug/blowdryer/Blowdryer.java
@@ -101,11 +101,14 @@ public class Blowdryer {
 				if (metaFile.exists() && dataFile.exists()) {
 					Map<String, String> props = loadPropertyFile(metaFile);
 					String propUrl = props.get(PROP_URL);
+					if (propUrl == null) {
+						throw new IllegalArgumentException("Unexpected content, recommend deleting file at " + metaFile);
+					}
 					if (propUrl.equals(url)) {
 						urlToContent.put(url, dataFile);
 						return dataFile;
 					} else {
-						throw new IllegalStateException("Expected url " + url + " but was " + propUrl + " for " + metaFile.getAbsolutePath());
+						throw new IllegalStateException("Expected url " + url + " but was " + propUrl + ", recommend deleting file at " + metaFile.getAbsolutePath());
 					}
 				} else {
 					Files.createParentDirs(dataFile);
@@ -191,6 +194,7 @@ public class Blowdryer {
 
 	/** Returns either the filename safe URL, or (first40)--(Base64 filenamesafe)(last40). */
 	static String filenameSafe(String url) {
+		url = preserveFileExtensionBitbucket(url);
 		String allSafeCharacters = url.replaceAll("[^a-zA-Z0-9-+_.]", "-");
 		String noDuplicateDash = allSafeCharacters.replaceAll("-+", "-");
 		if (noDuplicateDash.length() <= MAX_FILE_LENGTH) {
@@ -207,6 +211,19 @@ public class Blowdryer {
 					.replace('/', '-').replace('=', '-');
 			return first + "--" + hashed + end;
 		}
+	}
+
+	// preserve the filename and extension if query parameters are present in original url.
+	// required to retrieve XML files.
+	// From: https://mycompany.bitbucket.com/projects/PRJ/repos/my-repo/raw/src/main/resources/checkstyle/spotless.gradle?at=07f588e52eb0f31e596eab0228a5df7233a98a14
+	// To:   https://mycompany.bitbucket.com/projects/PRJ/repos/my-repo/raw/src/main/resources/checkstyle/spotless.gradle?at=07f588e52eb0f31e596eab0228a5df7233a98a14-spotless.gradle
+	private static String preserveFileExtensionBitbucket(String url) {
+		int atIdx = url.indexOf("?at=");
+		if (atIdx != -1) {
+			String fileNameWithoutQuery = url.substring(0, atIdx);
+			url = String.format("%s-%s", url, fileNameWithoutQuery.substring(fileNameWithoutQuery.lastIndexOf("/") + 1));
+		}
+		return url;
 	}
 
 	//////////////////////
@@ -253,7 +270,8 @@ public class Blowdryer {
 		}
 	}
 
-	static interface AuthPlugin {
+	@FunctionalInterface
+	interface AuthPlugin {
 		void addAuthToken(String url, Request.Builder builder) throws MalformedURLException;
 	}